+++ src/components/DatePicker/DatePicker.vue
@@ -126,17 +126,17 @@
 <script setup lang="ts">
 import { computed, onMounted } from 'vue'
 
-import Input from './Input.vue'
-import { Button } from './Button'
-import Popover from './Popover.vue'
-import FeatherIcon from './FeatherIcon.vue'
-import TextInput from './TextInput.vue'
+import Input from '../Input.vue'
+import { Button } from '../Button'
+import Popover from '../Popover.vue'
+import FeatherIcon from '../FeatherIcon.vue'
+import TextInput from '../TextInput.vue'
 
-import { getDate, getDateValue } from '../utils/dates'
-import { dayjsLocal } from '../utils/dayjs'
-import { useDatePicker } from '../utils/useDatePicker'
+import { getDate, getDateValue } from './utils'
+import { useDatePicker } from './useDatePicker'
+import { dayjsLocal } from '../../utils/dayjs'
 
-import type { DatePickerEmits, DatePickerProps } from './types/DatePicker'
+import type { DatePickerEmits, DatePickerProps } from './DatePicker'
 
 const props = defineProps<DatePickerProps>()
 const emit = defineEmits<DatePickerEmits>()
