+++ src/components/DatePicker/utils.ts
@@ -1,4 +1,4 @@
-import { dayjs } from './dayjs'
+import { dayjs } from '../../utils/dayjs'
 
 type DateConstructorParam = string | number | Date
