+++ src/index.js
@@ -7,9 +7,9 @@ export { default as Breadcrumbs } from './components/Breadcrumbs.vue'
 export * from './components/Button'
 export { default as Card } from './components/Card.vue'
 export { default as Checkbox } from './components/Checkbox.vue'
-export { default as DatePicker } from './components/DatePicker.vue'
-export { default as DateTimePicker } from './components/DateTimePicker.vue'
-export { default as DateRangePicker } from './components/DateRangePicker.vue'
+export { default as DatePicker } from './components/DatePicker/DatePicker.vue'
+export { default as DateTimePicker } from './components/DatePicker/DateTimePicker.vue'
+export { default as DateRangePicker } from './components/DatePicker/DateRangePicker.vue'
 export { default as Dialog } from './components/Dialog.vue'
 export { default as Dialogs } from './components/Dialogs.vue'
 export { default as Divider } from './components/Divider.vue'
